Made the MCMC an ABC!
This is more inline with what I had envisioned for this project. A user of this module defines different demographic models with minimal knowledge on what goes on behind the scenes. All one has to do now is define the parameters, the demographic model itself, and the model specific parameters associated with the MCMC. Need to finish documenting this all.