This should the MCMCB issues we were having earlier!
So... I was interpreting the results of the regression wrong. The first term is the **higher** order term, while the second is the constant. So the previous results were dependent on not where the intercept lied, but how steep or shallow this line really was. And I wasn't converting this back out of log scale either. Hopefully this gives us a better mixed result.